backend/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
psycopg2==2.9.3
